Test example works with web
