+++ examples/test/bin/test-mac/resources/css/subcss/test.css
@@ -1,0 +1,7 @@
+a {
+    color: deepskyblue;
+}
+
+a:hover {
+    color: deeppink;
+}
