+++ examples/test/menu.go
@@ -1,4 +1,4 @@
-package main
+package test
 
 import (
 	"net/url"
