--- examples/test/resources/css/subcss/test.css
@@ -1,7 +1,0 @@
-a {
-    color: deepskyblue;
-}
-
-a:hover {
-    color: deeppink;
-}
